+++ inventory_rest_api/Models/OrderPayment.cs
@@ -7,7 +7,7 @@ namespace inventory_rest_api.Models
         public long OrderPaymentId { get; set; }
         public long OrderSalesId { get; set; }
         public string PaymentOrderSalesDate { get; set; }
-        public long PaymentAmount { get; set; }
+        public double PaymentAmount { get; set; }
 
         [JsonIgnore]
         public OrderSales orderSales { get; set; }
