+++ inventory_rest_api/Models/PaymentPurchase.cs
@@ -7,7 +7,7 @@ namespace inventory_rest_api.Models
         public long PaymentPurchaseId { get; set; }
         public long PurchaseId { get; set; }
         public string PaymentPurchaseDate { get; set; }
-        public long PaymentAmount { get; set; }
+        public double PaymentAmount { get; set; }
 
         [JsonIgnore]
         public Purchase Purchase { get; set; }
